refactor(bitrix24): move Bitrix24Client to Infrastructure layer
Move the concrete HTTP implementation out of the Domain layer to
enforce DDD boundaries. The interface remains in Domain.
