+++ backend/app/Providers/Bitrix24ServiceProvider.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Providers;
 
 use App\Domain\Settings\Models\Setting;
-use App\Domain\Bitrix24\Services\Bitrix24Client;
+use App\Infrastructure\Bitrix24\Bitrix24Client;
 use App\Domain\Bitrix24\Services\Bitrix24ClientInterface;
 use Illuminate\Support\ServiceProvider;
